+++ src/main/java/org/c4marathon/assignment/usecase/MemberAccountUseCase.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.usecase;
 
 import org.c4marathon.assignment.domain.model.Member;
 import org.c4marathon.assignment.dto.member.MemberRegistrationRequestDto;
-import org.c4marathon.assignment.domain.service.MainAccountService;
+import org.c4marathon.assignment.domain.service.mainaccount.MainAccountService;
 import org.c4marathon.assignment.domain.service.MemberService;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.stereotype.Service;
